Reorganized folder paths to improve visibilty.
